+++ libbpf-cargo/src/build.rs
@@ -223,13 +223,18 @@ fn compile_one(
     strip_dwarf_info(out).with_context(|| format!("Failed to strip object file {}", out.display()))
 }
 
-fn compile(debug: bool, objs: &[UnprocessedObj], clang: &Path, target_dir: &Path) -> Result<()> {
+fn compile(
+    debug: bool,
+    objs: &[UnprocessedObj],
+    clang: &Path,
+    mut clang_args: Vec<OsString>,
+    target_dir: &Path,
+) -> Result<()> {
     let header_dir = extract_libbpf_headers_to_disk(target_dir)?;
-    let compiler_options = if let Some(dir) = &header_dir {
-        vec![OsString::from(format!("-I{}", dir.to_str().unwrap()))]
-    } else {
-        Vec::new()
-    };
+    if let Some(dir) = header_dir {
+        clang_args.push(OsString::from("-I"));
+        clang_args.push(dir.into_os_string());
+    }
 
     for obj in objs {
         let stem = obj.path.file_stem().with_context(|| {
@@ -245,7 +250,7 @@ fn compile(debug: bool, objs: &[UnprocessedObj], clang: &Path, target_dir: &Path
         let mut dest_path = obj.out.to_path_buf();
         dest_path.push(&dest_name);
         fs::create_dir_all(&obj.out)?;
-        compile_one(debug, &obj.path, &dest_path, clang, &compiler_options)?;
+        compile_one(debug, &obj.path, &dest_path, clang, &clang_args)?;
     }
 
     Ok(())
@@ -259,10 +264,12 @@ fn extract_clang_or_default(clang: Option<&PathBuf>) -> PathBuf {
     }
 }
 
+#[allow(clippy::too_many_arguments)]
 pub fn build(
     debug: bool,
     manifest_path: Option<&PathBuf>,
     clang: Option<&PathBuf>,
+    clang_args: Vec<OsString>,
     skip_clang_version_checks: bool,
 ) -> Result<()> {
     let (target_dir, to_compile) = metadata::get(debug, manifest_path)?;
@@ -281,7 +288,8 @@ pub fn build(
     let clang = extract_clang_or_default(clang);
     check_clang(debug, &clang, skip_clang_version_checks)
         .with_context(|| anyhow!("{} is invalid", clang.display()))?;
-    compile(debug, &to_compile, &clang, &target_dir).context("Failed to compile progs")?;
+    compile(debug, &to_compile, &clang, clang_args, &target_dir)
+        .context("Failed to compile progs")?;
 
     Ok(())
 }