+++ libbpf-cargo/CHANGELOG.md
@@ -4,6 +4,7 @@ Unreleased
   feature
 - Adjusted `SkeletonBuilder::clang_args` to accept an iterator of
   arguments instead of a string
+- Added `--clang-args` argument to `make` and `build` sub-commands
 - Updated `libbpf-sys` dependency to `1.3.0`
 - Bumped minimum Rust version to `1.70`